A couple of renames and code relocation to make even more sense
